backend/requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
pydantic-settings
qdrant-client
httpx
python-Levenshtein
langchain-groq
langchain-core
python-dotenv
